e1000: CSB acquire/release protocol revised
When writing CSBAH/CSBAL registers, the guest is always required to
write CSBAL last. When this happens, the host unmaps (releases) the
previously mapped CSB, if any. In the last case it also cancels the
tx_bh scheduling, so that no invalid CSB access can be done.

After that, if the new CSBAL/CSBAH content specifies a non-zero
physical address, the host maps (acquires) the new CSB.

==============================================================

The initialization of s->tx_count has been moved from start_xmit() to
e1000_tx_bh(), in order to avoid BH spinning in case the TX operation
is disabled in hardware with the TCTL register (see the if statement
at the beginning of start_xmit()).