rtl8139cp: paravirtual interrupt moderation
In paravirtual mode, a moderation timer has been implemented using
a QEMUTimer, and the register IntrMitigate, previously unimplemented.
Since I don't have the documentation for this register, I hust make up
a moderation specification: The register value DELTA is interpreted as a
moderation delay (in microseconds). If DELTA is zero, the moderation
is disabled. Otherwise the emulation makes sure that two consecutive
interrupts are never closer than DELTA microseconds.