rtl8139cp: guest_request_txkick
When the guest writes a valid TX descriptor index ([0, TX_RING_SIZE-1])
to csb->guest_request_txkick, the host will send an interrupt when
it processes that descriptor. The guest can use this feature in place
of guest_need_txkick (which should be always zero) in order to implement a
delayed form of "guest_need_txkick <= 1". This is useful when the guest
is faster than the host on the TX path and TX interrupts are used only when
the guest runs out of free TX descriptors.