e1000: host_need_rxkick_at
host_need_rxkick_at is used instead of host_need_txkick, so that
when the input load is very high, we minimize the RDT write rate:
Since the master/producer (QEMU e1000 RX frontend) is faster than
the slave/consumer (guest e1000 driver), back-notifications
(consumer-->producer notification due to more free space made
available to the producer by the consumer) are very frequent
with the "host_need_rxkick" mechanism, because the consumer
back-notifies the producer as soon as it has freed one slot.
Using the "host_need_rxkick" means back-notify only when the
consumer has freed most of the slot it can free.