e1000/rtl8139cp: got rid of host_txcycles(_lim)
host_txcycles has been removed CSB (it was useless), and
host_txcycles_lim is cached in an emulator private variable,
refreshed only at CSB configuration-time (so that for now if the
guest dinamically changes csb->host_txcycles_lim, the change has
no effect).

In this way we avoid accessing the CSB for the two variables.