e1000: mapping guest RX buffers
Instead of using pci_dma_write() (which performs an address
translation), we use the map_mbufs() function to get an host
virtual address and then memcpy() into the buffers. In this way
we avoid (most of the times) the address translation.

Also a redundant E1000_CONFIG_PARAVIRT block has been removed
in the e1000_receive() function.