e1000: vnet-hdr-unaware backends work with TSOv4
The e1000 driver always uses the virtio-net header, even if the
QEMU backend does not support it. This makes driver modifications
simpler. The frontend has to push a zero header in these cases.

With TSO packets we also have to zero the IP checksum, since the
driver doesn't do it anymore in paravirtual mode (e.g. in Linux,
this was done by e1000_tso()) and the e1000 specification
requires it, i.e. the e1000 emulation expects that field to be
zero before the checksumming computation starts.
This is not done anymore into the driver because in paravirtual
mode, assuming virtio-net header is ALWAYS used, we must keep the
IP checksum (correctly computed by the kernel), otherwise we have
to recompute it also for virtio-net-header-capable backends.