e1000: tcpudp_magic() done in the frontend
For TSO packets, a partial checksum computation is performed
by the frontend in paravirtual mode (when the backend does not
support the virtio-net header). The partial checksum includes the IP source
and destination addresses and the IP protocol field of the TSO
packet. The rest of the checksum (TCP segment length, TCP header, TCP
payload + final 1's complement) is also done by the emulator (the
segmentation code adds the length, while putsum() adds the TPC header
& payload, folds and complements).