e1000: remap_rings()
Remaps the TX and RX rings into the host virtual
memory, so that start_xmit and e1000_receive_iov can access them.
This function is called in both set_dba() and set_dlen(), so that
the mapping is correctly set up either when the guest first writes
to TDLEN (RDLEN) and when it first writes to TDBAH/TDBAL
(RDBAH/RDBAL).